I had a configuration error which was preventing these commit messages from being emailed to everyone.  That is fixed now, so you will see these.
I am checking in the plusAssoc proof we did today.  I also fixed the
bug with erasure of theta-terms, so now we can define plus using a
theta, and nat-thms is not broken by that.  Also, I am using theta in
some more places in nat-thms, and adding some more information for
theta-terms (namely, the expected type) to the info window.



git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/lambda-encodings/implementations/cedille@22571 bd178f5f-7bea-4014-844f-495ca19b8805