Together with an update to the IAL which I just committed (so you need to update your IAL, too), this seems to have fixed the problem I was debugging with your help this afternoon.
Speed-ups are not quite as dramatic as I would like, because currently
cedille is going to re-parse the file even if the file is not out of
date.  The issue is that cedille needs to know which imported files to
check, because if any of them is out of date, then we need to
re-process the file.  

One solution to this would be to store a list of the imported files in
another file (or possibly even scrape it out of the .cede file?), so
that cedille does not have to waste time re-parsing the file just to
find the imports.  Of course, another solution is to have a better
parser, but this will have to wait a bit more.

For now, I am seeing some modest speed-ups on larger files.  So this
can still help.



git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/lambda-encodings/implementations/cedille@22796 bd178f5f-7bea-4014-844f-495ca19b8805