I have incorporated Richard's changes, for F and B.  These let you move forward/backward if you are stuck with f/b.
Also included are changes so that if you jump to a symbol
defined in the same file, the selection will get cleared.
This looks a lot nicer.

I added support, too, for not trying to resize the info
window if it cannot be vertically resized (eg, if your
emacs pops up the info window with a vertical split instead
of a horizontal one).

BTW, you can add the follow to your .emacs file to avoid
vertical splits:

  (setq split-width-threshold nil)




git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/lambda-encodings/implementations/cedille@23187 bd178f5f-7bea-4014-844f-495ca19b8805