Using the built in system to highlight entire spans based on their 'name' entry. se-highlight contains a few defined faces and a [font -> list of names] map. cedille-mode simply has a hook that tells cedille to call 'se-highlight' after initializing spans. Also bound 'se-highlight' to the key 'C-h' but this can be removed.
Future -- it is straightforward to have multiple highlighting modes, and we could bind something like C-h 1, C-h 2, etc. to switch between them.

In the mid-term, it would be good to have a separate entry to annotate spans more carefully with information useful to highlighting and to cue off of that entry rather than 'name'.

In the long-term, we should probably move to a highlighting system like agda's that takes advantage of built in emacs customization, etc.  It would also be beneficial to move away from highlighting entire spans.




git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/lambda-encodings/implementations/cedille@23221 bd178f5f-7bea-4014-844f-495ca19b8805