This cedille-mode.el and cedille-mode folder have been adapted so that highlighting and errors are factored out and placed in the cedille-mode directory.
Instead of adding a new line to cedille-mode.el for each additional file in the cedille-mode file, I'm loading cedille-mode-library.el in cedille-mode.el and having it reference all of the files that need to be included.

My thought is that in the long-term we might increase dependecy on cedille-mode-library.el and decrese dependency on se.el and se-mode.el.
There may be better ways to implement this, but I figured I would go ahead and get the code factored.



git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/lambda-encodings/implementations/cedille@23255 bd178f5f-7bea-4014-844f-495ca19b8805