Add the info pages functionality to cedille-mode
The key binding for bringing up the info pages needs to be put in
cedille-mode.el (the line is in a comment at the top of
cedille-mode-info.el).

Docs now has the src files, but I omitted the compiled documentation
file.  Otherwise, everytime someone re-creates the info pages it
will show up as "new" whether it has changed or not.

The compile-docs.sh script is the recommended way to create
the docs, since it puts files in the correct place and with the
name that cedille-mode is looking for.  It adds the .info extension
to the help page to make it easier to ignore if we end up with multiple
help pages in the future.

The .keep file only exists to force git to track the empty info
directory for me.

git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/lambda-encodings/implementations/cedille@23271 bd178f5f-7bea-4014-844f-495ca19b8805