Add first two pages for cedille docs
When adding new pages the menu entry to the left of the double-colon
MUST match the name of the @node label on that page.  The @chapter
label can be anything, and the included filename can be anything.

I am not sure what the syntax is for sub-pages that will not be
included on the menu, but I will look into it and add an example
at some point over the next couple of days.

git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/lambda-encodings/implementations/cedille@23342 bd178f5f-7bea-4014-844f-495ca19b8805