Checking π proof-terms now will put the sides of the equation in question into head-normal form.  This means you no longer have to write π ε x.  Just π x will have that same effect.
I also did a little clean up in cedille-mode.el.  We used to have
(if x (prog y1 ... yk) nil).  It is nicer elisp to write instead

(when x y1 ... yk)

I also factored out the calls to show the info buffer and the
context into a helper function cedille-mode-show-extra-buffers.



git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/lambda-encodings/implementations/cedille@23357 bd178f5f-7bea-4014-844f-495ca19b8805