I added a new form of rewrite term, with this syntax:
ρ+ t - t'

This works like our existing ρ-term, except that it will
beta-reduce the type in question top-down, looking for
opportunities to rewrite t1 to t2 (where t proves t1 ≃ t2).

There are some cases where this does too much rewriting, so
one may find it necessary to use ρ sometimes instead of ρ+.

In cedille-mode I temporarily commented out the code which is
redisplaying the context buffer.  With this code commented out, my
info buffer does not resize now to 3 lines as we were seeing
Wednesday.  The context buffer of course can still be displayed,
but it will not be automatically updated when navigating.  This
change can be reverted once Ananda has modified the way the info
buffer is displayed, to match what we are doing for the context
and summary buffers.



git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/lambda-encodings/implementations/cedille@23365 bd178f5f-7bea-4014-844f-495ca19b8805