I am filtering out a few more attributes now if cedille-mode-debug is nil.
Also, I corrected one small issue where we would not generate spans for
the second part of a rho-term if we cannot synthesize a type for the
first part (now we do produce spans for the second part in that case).

I am removing the old src-0.9 which we do not need and are not using at all.

Finally, I got yet another version of the trm datatype working, in
alttrm/trmn.ced (this has taken a month or more off and on).



git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/lambda-encodings/implementations/cedille@23430 bd178f5f-7bea-4014-844f-495ca19b8805