Add cedille-find functionality to cedille-mode
-@ will find occurrences of highlighted symbol
-D will bring up the debug information

Still need to add some of the changes discussed in the last
meeting such as filtering by symbol occurrences in kinds, terms,
and types and also add jumping to the find buffer.

Backend changes to accomodate all of this.
Including updating putStr instances to putStrLn, except for when
spans are printed out because of an extra newline I could not find.
It might be from writeFile in the IAL?

git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/lambda-encodings/implementations/cedille@23937 bd178f5f-7bea-4014-844f-495ca19b8805