Fix Richard's bug where spans were not output
The single putStr instance was the issue for him, there was no
extra newline on his machine.

However, on my machine this generates the spurious communication bug.
If you receive that bug as I do, change line 223 in src/main.agda
from "putStrLn (include-elt-spans-to-string ie)" to use putStr.

I am not sure what the difference is, please test it out and see
if it works on your machines.

git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/lambda-encodings/implementations/cedille@23941 bd178f5f-7bea-4014-844f-495ca19b8805