Ernesto has volunteered to work on the Cedille parser, and so I cleaned up the grammar a bit, removing a few Greek notations no one is using (lowercase pi, delta, and omega).
Also, I noticed for some time now that some very old code in there is
automatically instantiating iota-types.  Sometimes this is cool,
because you do not have to write x.2 in a few cases.  Also, sometimes
you do not have to write [ t1 , t2 ].  But this auto-instantiation can
be confusing because the error messages can be a bit unexpected.  So I
removed that code.  It means that if you are using dependent
intersection types directly, you may find that some of your code needs
to have .2 explicitly applied to something.  Also, I found in my code
a couple of cases where I had to explicitly build a term of the form [
t1 , t2 ], where previously I did not have to do that.  The automatic
instantiation that was letting us avoid writing [ t1 , t2 ] in a few
cases may have been incorrect; I am not sure.  So you may want to
check out your developments carefully, when you update to this
version.

Finally, I added module declarations to the files in lib.



git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/lambda-encodings/implementations/cedille@28638 bd178f5f-7bea-4014-844f-495ca19b8805