If you load mini.ced with "no-cede-files" everything is fine in that file (there is a hole in imported file ri.ced).
But if you load it with the default option of producing cede
files, then at first there is no problem.  But if you edit
mini.ced and reload the file, you will see an error for the
body of last definition (of "Trm").  I am thinking this is
some kind of module qualification issue when we are reading
files without type checking them.



git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/lambda-encodings/implementations/cedille@29468 bd178f5f-7bea-4014-844f-495ca19b8805