This captures what seems to be a bug with module parameters.
The file a.ced declares a module with a parameter
F : * -> * -> *.

Then b.ced imports that module with a particular instantiation
for F.  A type error that seems wrong ensues.  Combining a and b
into one file ab.ced removes the error.

I am hoping Colin and/or Larry can take a look at this.



git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/lambda-encodings/implementations/cedille@29529 bd178f5f-7bea-4014-844f-495ca19b8805