replace INSTALL.txt with BUILD.md
to avoid maintaining duplicate documentation in this
repo (as INSTALL.txt) and the cedille.github.io repo
(as BUILD.md), I'm deleting INSTALL.txt and
replacing it with BUILD.md. Now we should consider
BUILD.md to be the canonical documentation for building,
and BUILD.md in cedille.github.io is a snapshot of this
file that is appropriate for the most recent cedille
release.