Changed cedille-archive-read-file-when-compile
I am having trouble byte-compiling some of the elisp, using emacs
25.2.2 (which I find on Ubuntu 18).  The code in
cedille-mode-archive.el that is supposed to read in the template files
for html/js emission seems not to work correctly for me when not
byte-compiled.  So I rewrote this code just using cedille-path-el to
get the path to the templates.  Sorry, @CarlOlson, if I misunderstood
your intent with that code.  Feel free to revert if I messed something
up.