issue #50: are-free-in-term now considers erasure occurrence of let bound var
in the case where are checking whether a variable occurs in a
Let binding, and we are skipping erased terms, if we find that
the var occurs in the definition body of the Let we don't
necessarily report that the var occurs. if we find that the
Let bound variable does occur then we report that the var does
occur.