Starting an encodings directory
When schematic Cedille is designed and implemented, we will change
over from the templates Colin has put in place, to use
telescope-generic versions of our encodings.  This is the start of
such, where I just pulled the functor into a module parameter.  I also
added a proof that inAlg is an initial algebra (Initiality in
Ind.ced).  As expected, the proof uses induction.