Correct definition of homomorphism of Mendler algebras
Thanks to Chris for pointing out that my definition was not following
that of the Uustalu Vene paper and to Stephan for helping me see that
the definition in that paper is correct.  I proved that we have an
initial algebra, using that notion.  I still would like to prove that
homomorphisms compose and that the identity function is a
homomorphism.  It would also be nice to derive some of the basic
consequences of initiality like fusion and (is it called?)
cancellation.