Yet another take on the notion of algebra
This one gives us a category of algebras with an initial object.  It
also makes algebraic sense, since it defines homomorphisms as
functions h where, in crude essence, h composed with alg1 equals alg2
composed with h.  So the homomorphism commutes past the operations of
the algebra, as expected from basic abstract algebra.