Rename the pair constructor to sigma, then define pair
The motiviation for this change is so that our type inference works
better in the common situation of building Pairs (non-dependent
pairs).  With sigma, one generally has to supply the type B : A ➔ ★,
forcing one also to supply A.  With a special constructor called pair
for non-dependent pairs, one need not supply A and B if the components
of the pair are synthesizing.