rename maybe-or to ||-maybe
The rationale for this change is to make it easier to parse
expressions in the code using this infix binary operator.  I was
seeing things like

blah yak maybe-or yak yak

and it took a while to realize that the maybe-or was the outermost
operator; it looked like it could be another argument to blah.