Logging functions factored out now into this module
I changed logMsg (which used to be in main.agda) so that it now needs
the log file path.  Previously it would recompute the log file path
every time a message was logged!  Yipes.  I then factored this
function and a couple related ones into this new module, so they can
be conveniently used in several places.