cleaned up how the toplevel-state record is used, also logging
Logging has been pulled into communication-util.agda, and requires now
a toplevel state, which necessitated propagating that toplevel state a
little deeper in the call graph than previously (to reach calls to
logMsg).

Also, we were unnecessarily matching on the toplevel state record in a
number of cases of process-cmd, and then rebuilding that state.
Agda's record notation gives a lighter way to do that.