removed maybe-if
The code was using a point-free idiom where we convert a boolean to a
maybe unit.  This led to some slightly inefficient and unnecessarily
tricky code, so I removed all calls to maybe-if, and its definition.
This was code cleanup (as well as an exercise in understanding the
point-free code we have).