fix bug in positivity checker, add debug output when check fails
There was a renaming bug when the positivity checker (in datatype-util)
checked an application of another datatype constructor.  At that
point, the positivity checker unfolds the definition of the datatype
as if it were Church-encoded, and checks positivity.  Nested
applications of the same datatype were not working because of a
renaming issue.  This commits fixes this, and closes #143.